+++ packages/react-components/src/CopyButton/index.tsx
@@ -1,20 +1,17 @@
 // Copyright 2017-2021 @polkadot/react-components authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { IconName } from '@fortawesome/fontawesome-svg-core';
-
 import React, { useCallback, useContext } from 'react';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import styled from 'styled-components';
 
-import StatusContext from './Status/Context';
-import Button from './Button';
-import { useTranslation } from './translate';
+import StatusContext from '../Status/Context';
+import { useTranslation } from '../translate';
+import Icon from './copyIcon.svg';
 
 interface Props {
   children?: React.ReactNode;
   className?: string;
-  icon?: IconName;
   label?: React.ReactNode;
   type?: string;
   isMnemonic?: boolean;
@@ -23,7 +20,7 @@ interface Props {
 
 const NOOP = () => undefined;
 
-function CopyButton ({ children, className = '', icon = 'copy', label, type, value }: Props): React.ReactElement<Props> {
+function CopyButton ({ children, className = '', type, value }: Props): React.ReactElement<Props> {
   const { t } = useTranslation();
   const { queueAction } = useContext(StatusContext);
 
@@ -46,23 +43,17 @@ function CopyButton ({ children, className = '', icon = 'copy', label, type, val
       >
         <div className='copyContainer'>
           {children}
-          <span className='copySpan'>
-            <Button
-              className='icon-button show-on-hover'
-              icon={icon}
-              isDisabled={!value}
-              label={label}
-              onClick={NOOP}
-            />
-          </span>
+          <div
+            className='icon-button show-on-hover'
+            onClick={NOOP}
+          >
+            <img alt='copy'
+              src={Icon as string}/>
+          </div>
         </div>
       </CopyToClipboard>
     </div>
   );
 }
 
-export default React.memo(styled(CopyButton)`
-  .copySpan {
-    white-space: nowrap;
-  }
-`);
+export default React.memo(CopyButton);