+++ packages/react-components/src/Modal/index.tsx
@@ -10,6 +10,7 @@ import React, { useContext } from 'react';
 import { Modal as SUIModal } from 'semantic-ui-react';
 import { ThemeContext } from 'styled-components';
 
+import CloseIcon from './images/close-icon.svg';
 import Actions from './Actions';
 import Column from './Column';
 import Columns from './Columns';
@@ -25,7 +26,7 @@ type ModalType = React.FC<ModalProps> & {
 
 function ModalBase (props: ModalProps): React.ReactElement<ModalProps> {
   const { theme } = useContext<ThemeDef>(ThemeContext);
-  const { children, className = '', header, open = true } = props;
+  const { children, className = '', header, onCancel, open = true } = props;
 
   return (
     <SUIModal
@@ -35,7 +36,14 @@ function ModalBase (props: ModalProps): React.ReactElement<ModalProps> {
       open={open}
     >
       {header && (
-        <SUIModal.Header>{header}</SUIModal.Header>
+        <SUIModal.Header>{header}
+          <div className='close-btn'
+            onClick={onCancel}>
+            <img alt='close'
+              src={CloseIcon as string}
+            />
+          </div>
+        </SUIModal.Header>
       )}
       {children}
     </SUIModal>