+++ packages/react-components/src/Static.tsx
@@ -33,7 +33,7 @@ function Static ({ children, className = '', defaultValue, help, isFull, isHidde
       label={label}
       withLabel={withLabel}
     >
-      <div className='ui--Static ui dropdown selection disabled'>
+      <div className='step-3-texts'>
         {value || defaultValue}
         {children}
       </div>